requirements.txt
requests
python-dotenv
lxml
resend
asyncio
